pyproject.toml
[tool.poetry]
name = "db_connectors"
version = "0.1.0"
description = "all db connectors"
authors = ["Francisco Alfaro <francisco.alfaro@walmart.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8, <3.9"
pandas = "^1.4.1"
loguru = "^0.6.0"
pyodbc = "^4.0.32"
SQLAlchemy = "^1.4.31"
python-dotenv = "^0.20.0"
google-cloud-bigquery = "^3.2.0"
ibm-db = "^3.1.2"



[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"
Sphinx = "^5.0.2"
sphinx-rtd-theme = "^1.0.0"
sphinx-jsonschema = "^1.19.1"
sphinxcontrib-bibtex = "^2.4.2"
flake8 = "^4.0.1"
black = "^22.3.0"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"